requirements.txt
Flask-API
flask-script
flask-sqlalchemy
psycopg2-binary
flask-migrate
